src/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
apng==0.3.4
ase==3.25.0
async-timeout==5.0.1
atomate2==0.0.14
attrs==25.3.0
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
bcrypt==4.3.0
boto3==1.37.35
botocore==1.37.35
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
custodian==2024.6.24
cycler==0.12.1
deprecated==1.2.18
dnspython==2.7.0
docker-pycreds==0.4.0
e3nn==0.5.6
easygraphics==2.0.1.2
emmet-core==0.83.9
eval-type-backport==0.2.2
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
gitpython==3.1.44
h5py==3.13.0
idna==3.10
imageio==2.27.0
importlib-resources==6.5.2
isodate==0.7.2
jinja2==3.1.6
jmespath==1.0.1
jobflow==0.1.19
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
latexcodec==3.0.0
lightning-utilities==0.14.3
loguru==0.7.3
maggma==0.71.5
markupsafe==3.0.2
matplotlib==3.5.2
mattersim==1.1.2
mongomock==4.3.0
monty==2025.3.3
mp-api==0.43.0
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.4.3
narwhals==1.35.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.0.76
opt-einsum==3.4.0
opt-einsum-fx==0.1.4
orjson==3.10.16
packaging==24.2
palettable==3.3.3
pandas==2.2.3
paramiko==3.5.1
phonopy==2.38.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pybtex==0.24.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.8.1
pydash==8.0.5
pygame==2.6.1
pyjwt==2.10.1
pymatgen==2024.4.13
pymongo==4.10.1
pynacl==1.5.0
pyparsing==3.2.3
pyqt5==5.15.11
pyqt5-qt5==5.15.16
pyqt5-sip==12.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==1.8.4
pytz==2025.2
pywavelets==1.6.0
pyyaml==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
s3transfer==0.11.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.10.1
seekpath==2.1.0
sentinels==1.0.0
sentry-sdk==2.26.1
setproctitle==1.3.5
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spglib==2.6.0
sshtunnel==0.4.0
symfc==1.3.4
sympy==1.13.3
tabulate==0.9.0
tensorboardx==2.6.2.2
threadpoolctl==3.6.0
tifffile==2024.8.30
torch-ema==0.3
torch-geometric==2.6.1
torch-runstats==0.2.0
torchmetrics==1.7.1
torchsummary==1.5.1
tqdm==4.61.2
triton==3.0.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uncertainties==3.2.2
urllib3==1.26.20
wandb==0.19.9
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
